WIP: make PageReloadOverlay a catch all
For this approach a secondary middleware layer is needed which executes
after all regular middleware is done to see if 'recoverable' flag was
not set by any of the primary middlewares. In this commit it was done by
adding '_elements2' and 'register2' to MiddlewareRegistry, but it's just
the simpliest thing required to achieve the goal and is to be discussed.

The page reload feature is caching an action that caused a
"fatal error". If the page reload dialog is canceled then it reemits
the fatal action with recoverable flag set to false (and this time
the reload overlay does not interfere as it react only to
'recoverable' === undefined).